Finish README
<main>
Hey Everyone, Its OJKT here, im Learning code for the first time. a little nervous and unmotivated but I have a goal in mind. The goal is a NET BOT.. a net bot to automate some of the easier and simpler things I do on the web, such as constantly refresh the news for updates or reposting sucessful photos on my pintrest page.
</main>